+++ lib/gatekeeper_client.dart
@@ -3,4 +3,5 @@ library;
 
 export 'gatekeeper.dart';
 export 'src/gatekeeper_client.dart';
+export 'src/gatekeeper_client_config.dart';
 export 'src/gatekeeper_const.dart';